[Refactor] #1301 Separated apply-magic-armors-base.cpp/h from apply-magic-armor.cpp/h
