+++ src/object-enchant/apply-magic-armor.h
@@ -1,11 +1,12 @@
 ﻿#pragma once
 
 #include "object-enchant/enchanter-base.h"
+#include "object-enchant/apply-magic-armors-base.h"
 #include "system/angband.h"
 
 struct object_type;
 struct player_type;
-class ArmorEnchanter : EnchanterBase {
+class ArmorEnchanter : ArmorEnchanterBase {
 public:
     ArmorEnchanter(player_type *owner_ptr, object_type *o_ptr, DEPTH level, int power);
     ArmorEnchanter() = delete;
@@ -19,9 +20,5 @@ protected:
     void give_cursed() override;
 
 private:
-    void set_ac_bonus();
     player_type *owner_ptr;
-    object_type *o_ptr;
-    DEPTH level;
-    int power;
 };
