+++ src/object-enchant/apply-magic-armors-base.h
@@ -1,0 +1,13 @@
+﻿#pragma once
+
+#include "object-enchant/enchanter-base.h"
+#include "system/angband.h"
+
+struct object_type;
+class ArmorEnchanterBase : EnchanterBase {
+protected:
+    ArmorEnchanterBase(object_type *o_ptr, DEPTH level, int power);
+    virtual ~ArmorEnchanterBase() = default;
+    object_type *o_ptr;
+    int power;
+};
